% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTCR.R
\name{categoryTCR}
\alias{categoryTCR}
\title{categoryTCR}
\usage{
categoryTCR(gene_list)
}
\arguments{
\item{gene_list}{list which is the result of CTgene split}
}
\value{
list contain vector of TCRA + TCRB of the cell input
}
\description{
change the format of splited CTgene
}
\keyword{internal}
