% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiation_probabilities.R
\name{differentiation_probabilities}
\alias{differentiation_probabilities}
\title{differentiation_entropy}
\usage{
differentiation_probabilities(
  wp_data,
  terminal_states = NULL,
  knn. = 30,
  n_jobs = -1,
  pseudotime,
  waypoints
)
}
\arguments{
\item{wp_data}{Multi scale data of the waypoints}

\item{knn.}{number of nearest neighbors for graph construction}

\item{n_jobs}{number of jobs for parallel processing}

\item{pseudotime}{pseudotime ordering of cells}

\item{use_early_cell_as_start}{bool}
}
\value{
entropy and probabilities
}
\description{
function to compute entropy and branch probabilities
}
