% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiationProbabilities.R
\name{differentiationProbabilities}
\alias{differentiationProbabilities}
\title{differentiationProbabilities}
\usage{
differentiationProbabilities(
  wp_data,
  terminal_states = NULL,
  knn. = 30L,
  pseudotime,
  waypoints
)
}
\arguments{
\item{wp_data}{Multi scale data of the waypoints}

\item{terminal_states}{integer, NULL by default. The index of the terminal state.}

\item{knn.}{integer, 30L by default. Number of nearest neighbors for graph construction.}

\item{pseudotime}{pseudotime ordering of cells}

\item{waypoints}{integer vector, index of selected waypoint used to construct markov chain}
}
\value{
probabilities
}
\description{
function to compute branch probabilities using markov chain
}
