% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.class.check}
\alias{.class.check}
\title{.class.check}
\usage{
.class.check(input, must)
}
\arguments{
\item{input}{the input need to be check}

\item{must}{the type we need}
}
\description{
check whether the input is with the correct class
}
