% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.classCheck}
\alias{.classCheck}
\title{.classCheck}
\usage{
.classCheck(input, must)
}
\arguments{
\item{input}{the input need to be check}

\item{must}{the type we need}
}
\value{
whether or not the input is the correct class
}
\description{
check whether the input is with the correct class
}
\keyword{internal}
