% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitCTgene.R
\name{.collapse_nested_list}
\alias{.collapse_nested_list}
\title{Collapse a nested list}
\usage{
.collapse_nested_list(input_list)
}
\arguments{
\item{input_list}{input nested list.}
}
\value{
collapsed list
}
\description{
Collapse a nested list
}
\keyword{internal}
