% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.markov.chain.R
\name{.construct.markov.chain}
\alias{.construct.markov.chain}
\title{.construct.markov.chain}
\usage{
.construct.markov.chain(wp_data, knn., pseudotime, waypoints)
}
\arguments{
\item{knn.}{Number of nearest neighbors for graph construction}

\item{pseudotime}{description}
}
\description{
markov chain construction
}
