% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determ.multiscale.space.R
\name{.determine.multiscale.space}
\alias{.determine.multiscale.space}
\title{.determine.multiscale.space}
\usage{
.determine.multiscale.space(diffusionmap, n_eigs = NULL)
}
\arguments{
\item{diffusionmap}{DiffusionMap object}

\item{n_eigs}{integer, default is NULL. Number of eigen vectors to use.
\itemize{
\item If is not specified, the number of eigen vectors will be determined using the eigen gap.
}}
}
\value{
dataframe
}
\description{
.determine.multiscale.space
}
