% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determMultiscaleSpace.R
\name{.determineMultiscaleSpace}
\alias{.determineMultiscaleSpace}
\title{.determineMultiscaleSpace}
\usage{
.determineMultiscaleSpace(diffusionmap, n_eigs = NULL)
}
\arguments{
\item{diffusionmap}{DiffusionMap object}

\item{n_eigs}{integer, default is NULL. Number of eigen vectors to use.
\itemize{
\item If is not specified, the number of eigen vectors will be determined using the eigen gap.
}}
}
\value{
dataframe
}
\description{
.determineMultiscaleSpace
}
\keyword{internal}
