% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalStateFromMarkovChain.R
\name{.determTerminal}
\alias{.determTerminal}
\title{.determTerminal}
\usage{
.determTerminal(terminal_states, i, dm_boudaries, wp_data)
}
\arguments{
\item{terminal_states}{integer vector to store the generated waypoint index}

\item{i}{iteration index}

\item{dm_boudaries}{index of the maxium or minium value of
transition matrix per row}

\item{wp_data}{Multi scale data of the waypoints}
}
\value{
integer vector store the index of waypoints serve as terminal state
}
\description{
function in Reduce to provide waypoints
}
\keyword{internal}
