% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVdjPseudobulk.R
\name{.extractVdj}
\alias{.extractVdj}
\title{Specify the columns which store VDJ information, and extract the main chain
from it}
\usage{
.extractVdj(sce, extract_cols, mode_option, verbose)
}
\arguments{
\item{sce}{SingleCellExperiment object input}

\item{extract_cols}{The setupVdjPseutobulk transfered parameter given by user
to specify the VDJ information columns}

\item{mode_option}{see document of setupVdjPseudobulk for
detailed explanation}

\item{verbose}{logical, print messages. Default is \code{TRUE}.}
}
\value{
SingleCellExperiment objects with column stores the information of
the main VDJ information in colData slot
}
\description{
Specify the columns which store VDJ information, and extract the main chain
from it
}
\keyword{internal}
