% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjPseudobulk.R
\name{.featureSpaceConstruct}
\alias{.featureSpaceConstruct}
\title{Construct VDJ feature space}
\usage{
.featureSpaceConstruct(milo, extract_cols, pbs)
}
\arguments{
\item{milo}{Milo or SingleCellExperiment object provided by user}

\item{extract_cols}{columns of names where to extract the VDJ information}

\item{pbs}{cell x pseudobulk adjacent matrix}
}
\value{
constructed feature space
}
\description{
Construct VDJ feature space
}
\keyword{internal}
