% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVdjPseudobulk.R
\name{.filterProductivity}
\alias{.filterProductivity}
\title{filer out cell with unproductive chain}
\usage{
.filterProductivity(
  sce,
  mode_option,
  productive_cols,
  productive_vj,
  productive_vdj,
  verbose
)
}
\arguments{
\item{sce}{SingleCellExperiment input}

\item{mode_option}{check setupVdjPseudobulk for detailed explanation}

\item{productive_vj}{If \code{TRUE}, retains cells where the main VJ chain is
productive.}

\item{productive_vdj}{If \code{TRUE}, retains cells where the main VDJ chain is
productive.}

\item{verbose}{logical, print messages. Default is \code{TRUE}.}
}
\value{
SingleCellExperiment object after filtering on producive chain
}
\description{
filer out cell with unproductive chain
}
\keyword{internal}
