% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVdjPseudobulk.R
\name{.filterUnmapped}
\alias{.filterUnmapped}
\title{Filter out cell with unclear mapping in VDJ information}
\usage{
.filterUnmapped(
  sce,
  mode_option,
  check_vj_mapping,
  check_vdj_mapping,
  main_cols,
  check_extract_cols_mapping,
  remove_missing
)
}
\arguments{
\item{sce}{SingleCellExperiment object input}

\item{mode_option}{see document of setupVdjPseudobulk for explanation}

\item{check_vj_mapping}{logical vector to set whether to check V and J gene
in VJ chain, passed from setupVdjPseudobulk}

\item{check_vdj_mapping}{logical vector to set whether to check
V, D and J gene in VDJ chain, passed from setupVdjPseudobulk}

\item{main_cols}{column names in colData in which The information of
main chain stores}

\item{check_extract_cols_mapping}{character vector,the names of columns that
needs to be checked, passed from setupVdjPseudobulk}

\item{remove_missing}{option for removing the unclear mappin or just mask it,
passed from setupVdjPseudobulk}
}
\value{
filtered SingleCellExperiment object
}
\description{
Filter out cell with unclear mapping in VDJ information
}
\keyword{internal}
