% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxMinSampling.R
\name{.findNewWaypoints}
\alias{.findNewWaypoints}
\title{function used in Reduce to find new waypoint in an iteration}
\usage{
.findNewWaypoints(iterdists, k, vecs, ind, datas)
}
\arguments{
\item{iterdists}{a list containing both waypoints deteted in
the former iterations and the distance matrix used to find waypoints}

\item{k}{the iteration number}

\item{vecs}{a numeric vector used to calculate distance of waypoints to
each points}

\item{ind}{colnames}
}
\value{
a list containing updated distance matrix and new waypoints
}
\description{
function used in Reduce to find new waypoint in an iteration
}
\keyword{internal}
