% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVdjPseudobulk.R
\name{.generateExtractColumn}
\alias{.generateExtractColumn}
\title{Check whether the columns with specified names exist, if not, create them
with CTgene columns}
\usage{
.generateExtractColumn(sce, extract_cols, verbose)
}
\arguments{
\item{sce}{SingleCellExperiment object input}

\item{extract_cols}{column names we aim to extract information from}

\item{verbose}{logical, print messages. Default is \code{TRUE}.}
}
\value{
SingleCellExperiment with columns containing VDJ information in the
names we've specified.
}
\description{
Check whether the columns with specified names exist, if not, create them
with CTgene columns
}
\keyword{internal}
