% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVdjPseudobulk.R
\name{.generateExtractName}
\alias{.generateExtractName}
\title{Generate the name of columns with given parameter}
\usage{
.generateExtractName(sce, mode_option)
}
\arguments{
\item{sce}{SingleCellExperiment object input}

\item{mode_option}{see document of setupVdjPseudobulk for explanation}
}
\value{
a vecotor of colnames we need to perform main chain extraction
}
\description{
Generate the name of columns with given parameter
}
\keyword{internal}
