% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pbs.R
\name{.get.pbs}
\alias{.get.pbs}
\title{.get.pbs}
\usage{
.get.pbs(pbs, obs_to_bulk, milo)
}
\description{
Helper function to ensure we have cells by pseudobulks matrix which we can use for pseudobulking. Use the pbs and obs_to_bulk inputs to vdj_pseudobulk() and gex_pseudobulk()
}
