% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPbs.R
\name{.getPbsCol}
\alias{.getPbsCol}
\title{.getPbsCol}
\usage{
.getPbsCol(pbs, col_to_take, milo)
}
\arguments{
\item{pbs}{dgeMatrix, cell x pseudobulk binary matrix}

\item{col_to_take}{character vector, names of colData of milo that need to be processed}

\item{milo}{Milo or SingleCellExperiment object}
}
\value{
pbs_col, a DataFrame which will be passed to the new SingleCellExperiment object as colData of vdj x pseudobulk assays
}
\description{
Helper function to create the new pseudobulk object's coldata.
}
\keyword{internal}
