% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPbs.R
\name{.getPbsPerCol}
\alias{.getPbsPerCol}
\title{.getPbsPerCol}
\usage{
.getPbsPerCol(pbs.col, anno_col, milo, pbs)
}
\arguments{
\item{pbs.col}{DataFrame object used to store the result of each iteration}

\item{anno_col}{colname where to generate the metadata from}

\item{milo}{milo or SingleCellExperiment objects provided by user}

\item{pbs}{dgeMatrix, cell x pseudobulk binary matrix}
}
\value{
DataFrame object, serve as part of metadata of the new milo object
}
\description{
function used in Reduce to get the PbsCol
}
\keyword{internal}
