% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMarkovChain.R
\name{.KNNind}
\alias{.KNNind}
\title{Calculate the weight adjacent matricks of knn graph and its index}
\usage{
.KNNind(wp_data, knn.)
}
\arguments{
\item{wp_data}{Multi scale data of the waypoints}

\item{knn.}{Number of nearest neighbors for graph construction}
}
\value{
a list containing the weight adjacent matrix and index
}
\description{
Calculate the weight adjacent matricks of knn graph and its index
}
\keyword{internal}
