% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.min.sampling.R
\name{.max.min.sampling}
\alias{.max.min.sampling}
\title{.max.min.sampling}
\usage{
.max.min.sampling(data, num_waypoints)
}
\arguments{
\item{data}{data matrix along which to sample the waypoints, usually diffusion components}

\item{num_waypoints}{number of waypoints to sample}
}
\value{
Series reprenting the sampled waypoints
}
\description{
function for max min sampling of waypoints
}
