% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minMaxScale.R
\name{.minMaxScale}
\alias{.minMaxScale}
\title{minMaxScale}
\usage{
.minMaxScale(data)
}
\arguments{
\item{data}{dataframe need to be scale}
}
\value{
scaled value
}
\description{
scale the value to range 0 to 1
}
\keyword{internal}
