% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjPseudobulk.R
\name{.normalizeFeatureSpace}
\alias{.normalizeFeatureSpace}
\title{Normalize Feature Space}
\usage{
.normalizeFeatureSpace(
  pseudo_vdj_feature,
  extract_cols,
  min_count,
  renormalize,
  milo
)
}
\arguments{
\item{pseudo_vdj_feature}{constructed feature space}

\item{extract_cols}{names of columns to extract the VDJ information}

\item{min_count}{the minim count of a V/D/J gene}

\item{renormalize}{Whether to renormalize the matrix}

\item{milo}{Milo or SingleCellExperiment object provided by user}
}
\value{
normalized VDJ feature space
}
\description{
Make sure the sum of each V, D, and J gene within a pseudobulk equals to 1
}
\keyword{internal}
