% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjPseudobulk.R
\name{.normalizePerVDJ}
\alias{.normalizePerVDJ}
\title{function to normalize a specific kind of VDJ gene in feature space}
\usage{
.normalizePerVDJ(
  pseudo_vdj_feature,
  col_n,
  renormalize,
  define.mask,
  milo,
  min_count
)
}
\arguments{
\item{pseudo_vdj_feature}{constructed feature space}

\item{col_n}{name of column to extract the VDJ information}

\item{renormalize}{Whether to renormalize the matrix}

\item{define.mask}{logical vector determine whether the V/D/J gene
should be masked when normalizing}

\item{milo}{Milo or SingleCellExperiment object provided by user}

\item{min_count}{the minim count of a V/D/J gene}
}
\value{
feature space normalized on specifed V/D/J gene
}
\description{
function to normalize a specific kind of VDJ gene in feature space
}
\keyword{internal}
