% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjPseudobulk.R
\name{.packFeatureSpace}
\alias{.packFeatureSpace}
\title{Pack the normalized featuer space into new Milo object}
\usage{
.packFeatureSpace(pbs, col_to_take, milo, pseudo_vdj_feature)
}
\arguments{
\item{pbs}{cell x pseudobulk adjacent matrix}

\item{col_to_take}{Optional character or vector of characters. Specifies names of colData of milo that need to identify the most common value for each pseudobulk}

\item{milo}{Milo or SingleCellExperiment object provided by user}

\item{pseudo_vdj_feature}{VDJ feature space}
}
\value{
Milo object with VDJ feature space stored in its assay
}
\description{
The metadata will derived from the original milo
}
\keyword{internal}
