% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMarkovChain.R
\name{.removeEdge}
\alias{.removeEdge}
\title{function used in Reduce to remove KNN's backward edges except for
edges that are within the computed standard deviation}
\usage{
.removeEdge(Knn, i, rem_edges)
}
\arguments{
\item{Knn}{weight KNN adjacent matrix}

\item{i}{the iteration number}

\item{rem_edges}{the edges that need to be removes}
}
\value{
an updated matrix after one round of iteration
}
\description{
function used in Reduce to remove KNN's backward edges except for
edges that are within the computed standard deviation
}
\keyword{internal}
