% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.state.from.markov.chain.R
\name{.terminal.state.from.markov.chain}
\alias{.terminal.state.from.markov.chain}
\title{.terminal.state.from.markov.chain}
\usage{
.terminal.state.from.markov.chain(Transmat, wp_data, pseudotime, waypoints)
}
\value{
terminal_state
}
\description{
.terminal.state.from.markov.chain
}
