% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalStateFromMarkovChain.R
\name{.terminalStateFromMarkovChain}
\alias{.terminalStateFromMarkovChain}
\title{Determine terminal states using Markov chain if end states are not provided.}
\usage{
.terminalStateFromMarkovChain(
  Transmat,
  wp_data,
  pseudotime,
  waypoints,
  verbose
)
}
\arguments{
\item{Transmat}{Transition matrix}

\item{wp_data}{Multi scale data of the waypoints}

\item{pseudotime}{numeric vector, pseudotime of each pseudobulk}

\item{waypoints}{integer vector, waypoint selected to construct markov chain.}

\item{verbose}{Boolean, whether to print messages/warnings.}
}
\value{
terminal_state
}
\description{
Determine terminal states using Markov chain if end states are not provided.
}
\keyword{internal}
