% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.type.check}
\alias{.type.check}
\title{.type.check}
\usage{
.type.check(input, must)
}
\arguments{
\item{input}{the input need to be check}

\item{must}{the type we need}
}
\description{
check whether the input has the correct type
}
