% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.typeCheck}
\alias{.typeCheck}
\title{.typeCheck}
\usage{
.typeCheck(input, must)
}
\arguments{
\item{input}{the input need to be check}

\item{must}{the type we need}
}
\value{
whether or not the input is the correct type
}
\description{
check whether the input has the correct type
}
\keyword{internal}
