% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxMinSampling.R
\name{.waypiontsPerCol}
\alias{.waypiontsPerCol}
\title{find the waypoints according to certain columns of data}
\usage{
.waypiontsPerCol(waypoints, ind, datas, no.iterations)
}
\arguments{
\item{waypoints}{integer vector used to store waypoints}

\item{ind}{columns' colnames}

\item{datas}{scaled diffusionmap}
}
\value{
a numeric vector containing waypoints' index
}
\description{
find the waypoints according to certain columns of data
}
\keyword{internal}
