% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitCTgene.R
\name{formatVdj}
\alias{formatVdj}
\title{Change the format of splitCTgene output.}
\usage{
formatVdj(gene_list)
}
\arguments{
\item{gene_list}{list containing the output from splitCTgene.}
}
\value{
list contain vector of VJ + VDJ information of the cell input
}
\description{
Change the format of splitCTgene output.
}
\keyword{internal}
