% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milo.umap.R
\name{milo.umap}
\alias{milo.umap}
\title{milo.umap}
\usage{
milo.umap(
  milo,
  slot_name = "UMAP_knngraph",
  n.neighbors = 50L,
  metric = "euclidean"
)
}
\arguments{
\item{milo}{the milo object with knn graph that needed to conduct umap on.}

\item{slot_name}{character, with default "UMAP_knngraph".
The slot name in reduceDim where the result store}

\item{n.neighbors}{integer, with default 50L.
the number of neighboring points used}

\item{metric}{character, the choice of metric used to measure distance}
}
\value{
milo object with umap reduction
}
\description{
use function RunUMAP from Seurat to conduct the umap on the knn graph in milo object
}
