% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milo_umap.R
\name{milo_umap}
\alias{milo_umap}
\title{milo_umap}
\usage{
milo_umap(
  milo,
  slot_name = "UMAP_knngraph",
  n.neighbors = 50L,
  metric = "euclidean"
)
}
\arguments{
\item{milo}{the milo object with knn graph that needed to conduct umap on.}

\item{slot_name}{character, with default "UMAP_knngraph".
\itemize{
\item The slot name in reduceDim where the result store
}}

\item{n.neighbors}{integer, with default 50L.
\itemize{
\item the number of neighboring points used
\item parameter of RunUMAP, checking its document for further detail
}}

\item{metric}{character, with default "euclidean"
\itemize{
\item the choice of metric used to measure distance
\item parameter of RunUMAP, checking its document for further detail
}}
}
\value{
milo object with umap reduction
}
\description{
use function RunUMAP from Seurat to conduct the umap on the adjacency matrix of knn graph in milo object
}
\examples{
# generate example milo data
ex <- matrix(runif(100000)*100,nrow=100)
exx <- SingleCellExperiment::SingleCellExperiment(
assay = list(counts = ex, logcounts=log1p(ex))) 
exmilo <- miloR::Milo(exx)
# caculate knn graph
exmilo<-scater::runPCA(exmilo, assay.type = "logcounts", ncomponents = 20L)
exmilo <- miloR:: buildGraph(exmilo, k = 10L, d =20L)
exmlo <- miloR::makeNhoods(exmilo, k=10L, d=20L)
rownames(colData(exmilo)) <- 1:ncol(exmilo)
exmilo <- milo_umap(exmilo, n.neighbors = 10L, metric = "euclidean")

}
