% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_probability.R
\name{project_probability}
\alias{project_probability}
\title{project_probabilty}
\usage{
project_probability(diffusionmap, waypoints, probabilities, t = 1, ...)
}
\arguments{
\item{diffusionmap}{diffusion map, used to reconstruct diffustion distance matrix}

\item{waypoints}{index of waypoints}

\item{probabilities}{waypoints' probabilities, result from markov chain}

\item{t}{diffusion time}
}
\value{
each pseudobulk's probabilites
}
\description{
project the probabilities from Markov chain to each pseudobulk
}
