% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pseudotime_to_cell.R
\name{project_pseudotime_to_cell}
\alias{project_pseudotime_to_cell}
\title{project_pseudotime_to_cell}
\usage{
project_pseudotime_to_cell(adata, pb_adata, term_states, suffix = "")
}
\arguments{
\item{adata}{SingleCellExperiment object}

\item{pb_adata}{pseudobulk data}

\item{term_states}{vector of terminal states with branch_probabilities to be transferred}

\item{suffix}{suffix to be added after the added column names, default ''}
}
\value{
subset of adata where cell that do not belong to any neighbourhood are removed and projected pseudotime information stored colData
}
\description{
Function to project pseudotime & branch probabilities from pb.milo (pseudobulk) to adata (cell).
}
