% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPseudotimeToCell.R
\name{project_single_value}
\alias{project_single_value}
\title{Function to project pseudobulk-level values to single-cell level}
\usage{
project_single_value(x, y, value_name, verbose = TRUE)
}
\arguments{
\item{x}{Numeric vector, pseudobulk-level value to be projected.}

\item{y}{Matrix (pseudobulk x cell), used to project x back to cell level.}

\item{value_name}{Character, name of the value being projected.}

\item{verbose}{Boolean, whether to print messages/warnings.}
}
\value{
Numeric vector of projected values at cell level.
}
\description{
Function to project pseudobulk-level values to single-cell level
}
