% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectProbability.R
\name{projectProbability}
\alias{projectProbability}
\title{projectProbability}
\usage{
projectProbability(diffusionmap, waypoints, probabilities, t = 1)
}
\arguments{
\item{diffusionmap}{diffusion map, used to reconstruct diffustion distance matrix}

\item{waypoints}{index of waypoints}

\item{probabilities}{waypoints' probabilities, result from markov chain}

\item{t}{diffusion time}
}
\value{
each pseudobulk's probabilites
}
\description{
project the probabilities from Markov chain to each pseudobulk
}
