% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vdj_pseudobulk.R
\name{setup_vdj_pseudobulk}
\alias{setup_vdj_pseudobulk}
\title{setup_vdj_pseudobulk}
\usage{
setup_vdj_pseudobulk(
  sce,
  mode_option = c("abT", "gdT", "B"),
  productive_cols = NULL,
  productive_vdj = TRUE,
  productive_vj = TRUE,
  subsetby = NULL,
  groups = NULL,
  allowed_chain_status = c("Single pair", "Extra pair", "Extra pair-exception",
    "Orphan VDJ", "Orphan VDJ-exception"),
  extract_cols = NULL,
  filter_pattern = ",|None|No_cotig",
  check_vdj_mapping = c("v_call", "j_call"),
  check_vj_mapping = c("v_call", "j_call"),
  check_extract_cols_mapping = NULL,
  remove_missing = TRUE,
  already.productive = TRUE,
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object,
vdj data should contained in colData for filtering}

\item{mode_option}{optional
mode for extraction the V(D)J genes.
\itemize{
\item for productive filtering, function will check colData(s) named productive_{mode_option}_{type}, where type should be "VDJ" or "VJ" or both, depending on productive_vj and productive_vdj.
\itemize{
\item If set as "NULL", the function needs the option "extract_cols" to be specified
}
\item for
}}

\item{productive_cols}{paramter used in productivity filtering, NULL by default
\itemize{
\item colData names where VDJ/VJ information is stored so that this will be used instead of the standard columns.
\item must be be specified when productivity filtering is need to conduct and mode_option is NULL.
}}

\item{productive_vdj}{logical, TRUE by default
Option in productivity filtering. If True, cell will only be kept if the main VDJ chain is productive}

\item{productive_vj}{logical, TRUE by default
Option in productivity filtering. If True, cell will only be kept if the main VJ chain is productive}

\item{subsetby}{character, NULL by default
\itemize{
\item name of one colData provided for subsetting. The subsetting process will only start when both groups and subsetby are provided.
\item If provided, only the cell with {subsetby} feature in {groups} will be used for computing the VDJ feature space
}}

\item{groups}{character vector, NULL by default
\itemize{
\item condition for subsetting
\item If provided, only the following groups/ categories will be used for computing the VDJ feature space
}}

\item{allowed_chain_status}{character vectors, optional
element should come from c("single pair","Extra pair", "Extra pair-exceptipn", "Orphan VDJ","Orphan VDJ-exception")}

\item{extract_cols}{character vector, NULL by default
column names where VDJ/VJ information is stored so that this will be used instead of the standard columns}

\item{filter_pattern}{character string, optional ",|None|No_contig" by default
pattern to filter form object. If NULL, does not filter}

\item{check_vdj_mapping}{character vector, optional
elements should come from c("v_call", "j_call")
Only columns in the argument will be checked for unclear mapping(contain comma) in VDJ calls}

\item{check_vj_mapping}{character vector, optional
elements should come from c("v_call", "j_call")
Only columns in the argument will be checked for unclear mapping(contain comma) in VJ calls}

\item{check_extract_cols_mapping}{character vecter, NULL by default
Only columns in the argument will be checked for unclear mapping (containing comma) in columns specified in extract_cols}

\item{remove_missing}{bool, True by default
If true, will remove cells with contigs matching the filter from the object. If False, will mask them with a uniform value dependent on the column name.}

\item{already.productive}{logical. Whether the data is already filtered by productivity.
\itemize{
\item If true, the function will skip the productivity filtering
\item If false, retain only productive entries in columns specified by mode_option or productive_cols
}}
}
\value{
filtered SingleCellExperiment object
}
\description{
filtering the data based on productivity, chain status, and subset data. ....
}
