% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTCR.R
\name{splitCTgene}
\alias{splitCTgene}
\title{split_CTgene}
\usage{
splitCTgene(combined_Expression)
}
\arguments{
\item{combined_Expression}{SingleCellExperiment object after conducting scRepertoire::combineTCR()}
}
\value{
list contain vector of vj & vdj of TCR from each cell
}
\description{
split the vdj gene from CTgene and store them separately in the columns
}
\keyword{internal}
