% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTCR.R
\name{vecAssign}
\alias{vecAssign}
\title{vecAssign}
\usage{
vecAssign(vec, num)
}
\arguments{
\item{vec}{list of vdj(vj) gene to assign to the reusult}

\item{num}{number of gene should return, should be 2(vj) or 3(vdj)}
}
\value{
list contain vector of TCRA + TCRB of the cell input
}
\description{
Assgin the vdj(vj) gene to the result
}
\keyword{internal}
