% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectProbability.R
\name{.calDif}
\alias{.calDif}
\title{function help to reconstruct diffusion distance using Reduce}
\usage{
.calDif(dfm, i, j, eigenvector, lambda_t, K)
}
\arguments{
\item{dfm}{an nrow(eigenvectors) x nrow(eigenvectors) matrix
need to be filled with diffusion distance with iteration}

\item{i}{integer the row selected in this iteration}

\item{j}{integer the col selected in this iteration}

\item{eigenvector}{numeric vector, the eigenvector from diffusion map}

\item{lambda_t}{eigenvalues to the power of t(diffusion time)}

\item{K}{The number of the eigenvectors to be used in calculation}
}
\value{
updated diffusion distance matrix after one iteration
}
\description{
function help to reconstruct diffusion distance using Reduce
}
\keyword{internal}
