% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.class.check}
\alias{.class.check}
\title{.class.check}
\usage{
.class.check(input, must)
}
\arguments{
\item{must}{the type we need}

\item{check}{the input need to be check}
}
\description{
check whether the input has the correct type
}
