% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.markov.chain.R
\name{.construct.markov.chain}
\alias{.construct.markov.chain}
\title{.construct.markov.chain}
\usage{
.construct.markov.chain(wp_data, knn., pseudotime, waypoints)
}
\arguments{
\item{wp_data}{Multi scale data of the waypoints}

\item{knn.}{Number of nearest neighbors for graph construction}

\item{pseudotime}{pseudotime ordering of cells}
}
\value{
transition matrix of the markov chain
}
\description{
markov chain construction
}
