% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMarkovChain.R
\name{.constructMarkovChain}
\alias{.constructMarkovChain}
\title{.constructMarkovChain}
\usage{
.constructMarkovChain(wp_data, knn., pseudotime, waypoints, verbose = TRUE)
}
\arguments{
\item{wp_data}{Multi scale data of the waypoints}

\item{knn.}{Number of nearest neighbors for graph construction}

\item{pseudotime}{pseudotime ordering of cells}

\item{waypoints}{integer vector, index of selected waypoint used to}

\item{verbose}{whether to print messages
construct markov chain}
}
\value{
transition matrix of the markov chain
}
\description{
Markov chain construction
}
\keyword{internal}
