% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMarkovChain.R
\name{.constructMarkovChain}
\alias{.constructMarkovChain}
\title{.constructMarkovChain}
\usage{
.constructMarkovChain(wp_data, knn., pseudotime, waypoints, vb, use_RANN)
}
\arguments{
\item{wp_data}{Multi scale data of the waypoints}

\item{knn.}{Number of nearest neighbors for graph construction}

\item{pseudotime}{pseudotime ordering of cells}

\item{waypoints}{integer vector, index of selected waypoint used to}

\item{vb}{whether to print messages}

\item{use_RANN}{parameter to make user choose
whether to use RANN to construct Markov chain,
or keep using bluster}
}
\value{
transition matrix of the markov chain
}
\description{
Markov chain construction
}
\keyword{internal}
