% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determ.multiscale.space.R
\name{.determine.multiscale.space}
\alias{.determine.multiscale.space}
\title{.determine.multiscale.space}
\usage{
.determine.multiscale.space(diffusionmap, n_eigs = NULL)
}
\arguments{
\item{diffusionmap}{DiffusionMap object}

\item{n_eigs}{integer
Number of eigen vectors to use.
If NULL is specified, the number of eigen vectors will be determined using the eigen gap. Default is NULL.}
}
\value{
dataframe
}
\description{
.determine.multiscale.space
}
