% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjPseudobulk.R
\name{.featureSpaceConstruct}
\alias{.featureSpaceConstruct}
\title{Construct VDJ feature space}
\usage{
.featureSpaceConstruct(milo, extract_cols, pbs)
}
\arguments{
\item{milo}{Milo or SingleCellExperiment object provided by user}

\item{extract_cols}{columns of names where to extract the VDJ information}

\item{pbs}{cell x pseudobulk adjacent matrix}
}
\description{
perform matrix multiplication of pseudobulks by cell matrix by a cells by
VJs matrix strat off by creating the cell by VJs matix skip the prefix
stuff as the VJ genes will be unique in the columns
}
\keyword{internal}
