% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pbs.R
\name{.get.pbs}
\alias{.get.pbs}
\title{.get.pbs}
\usage{
.get.pbs(pbs, col_to_bulk, milo)
}
\arguments{
\item{pbs}{pbs parameter provided by vdj_pseudobulk(),  cells by pseudobulks matrix or NULL}

\item{col_to_bulk}{col_to_bulk parameter provided by vdj_pseudobulk(), column's name of colData from milo}

\item{milo}{SingleCellExperiment object}
}
\value{
a cell x pseudobulk matrix
}
\description{
Helper function to ensure we have cells by pseudobulks matrix which we can use for pseudobulking.
}
