% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPbs.R
\name{.getPbs}
\alias{.getPbs}
\title{.getPbs}
\usage{
.getPbs(pbs, col_to_bulk, milo)
}
\arguments{
\item{pbs}{pbs parameter provided by vdjPseudobulk(),  cells by pseudobulks matrix or NULL}

\item{col_to_bulk}{col_to_bulk parameter provided by vdjPseudobulk(), column's name of colData from milo}

\item{milo}{SingleCellExperiment object}
}
\value{
a cell x pseudobulk matrix
}
\description{
Helper function to ensure we have cells by pseudobulks matrix which we can use for pseudobulking.
}
\keyword{internal}
