% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmaxScale.R
\name{.minmaxScale}
\alias{.minmaxScale}
\title{minmaxScale}
\usage{
.minmaxScale(data)
}
\arguments{
\item{data}{dataframe need to be scale}
}
\value{
scaled value
}
\description{
scale the value to range 0 to 1
}
