% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjPseudobulk.R
\name{.normalizeFeatureSpace}
\alias{.normalizeFeatureSpace}
\title{Normalize Feature Space, make sure the sum of each V, D, and J gene within
a pseudobulk equals to 1}
\usage{
.normalizeFeatureSpace(
  pseudo_vdj_feature,
  extract_cols,
  min_count,
  renormalise,
  milo
)
}
\arguments{
\item{pseudo_vdj_feature}{constructed feature space, passed
from vdjPseudobulk}

\item{extract_cols}{names of columns to extract the VDJ information}

\item{min_count}{the minim count of a V/D/J gene}

\item{renormalise}{Whether to renormalise the matrix after set the gene
below min_count to 0}

\item{milo}{Milo or SingleCellExperiment object provided by user}
}
\value{
normalized VDJ feature space
}
\description{
Normalize Feature Space, make sure the sum of each V, D, and J gene within
a pseudobulk equals to 1
}
\keyword{internal}
