% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjPseudobulk.R
\name{.NormalizePerVDJ}
\alias{.NormalizePerVDJ}
\title{function to normalize a specific kind of VDJ gene in feature space}
\usage{
.NormalizePerVDJ(
  pseudo_vdj_feature,
  col_n,
  renormalise,
  define.mask,
  milo,
  min_count
)
}
\arguments{
\item{pseudo_vdj_feature}{constructed feature space, passed
from vdjPseudobulk}

\item{col_n}{name of column to extract the VDJ information}

\item{renormalise}{Whether to renormalise the matrix after set
the gene below min_count to 0}

\item{define.mask}{logical vector determin whether the V/D/J gene should
be mask when normalizing}

\item{milo}{Milo or SingleCellExperiment object provided by user}

\item{min_count}{the minim count of a V/D/J gene}
}
\value{
feature space normalized on specifed V/D/J gene
}
\description{
function to normalize a specific kind of VDJ gene in feature space
}
\keyword{internal}
