% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVdjPseudobulk.R
\name{.subsetSce}
\alias{.subsetSce}
\title{Subset sce with given parameter}
\usage{
.subsetSce(sce, subsetby, groups, verbose)
}
\arguments{
\item{sce}{SingleCellExperiment object input}

\item{subsetby}{subsetby Character. Name of a \code{colData} column for
subsetting. given by setupVdjPsudobulk.}

\item{groups}{Character vector. Specifies the subset condition for filtering.
given by setupVdjPsudobulk.}

\item{verbose}{logical, print messages. Default is \code{TRUE}.}
}
\value{
subsetted SingleCellExperiment object
}
\description{
Subset sce with given parameter
}
\keyword{internal}
