% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.state.from.markov.chain.R
\name{.terminal.state.from.markov.chain}
\alias{.terminal.state.from.markov.chain}
\title{.terminal.state.from.markov.chain}
\usage{
.terminal.state.from.markov.chain(Transmat, wp_data, pseudotime, waypoints)
}
\arguments{
\item{Transmat}{Transition matrix}

\item{wp_data}{Multi scale data of the waypoints}

\item{pseudotime}{numeric vector, pseudotime of each pseudobulk}

\item{waypoints}{integer vector, waypoint selected to construct markov chain.}
}
\value{
terminal_state
}
\description{
.terminal.state.from.markov.chain
}
