% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalStateFromMarkovChain.R
\name{.terminalStateFromMarkovChain}
\alias{.terminalStateFromMarkovChain}
\title{.terminalStateFromMarkovChain}
\usage{
.terminalStateFromMarkovChain(Transmat, wp_data, pseudotime, waypoints)
}
\arguments{
\item{Transmat}{Transition matrix}

\item{wp_data}{Multi scale data of the waypoints}

\item{pseudotime}{numeric vector, pseudotime of each pseudobulk}

\item{waypoints}{integer vector, waypoint selected to construct markov chain.}
}
\value{
terminal_state
}
\description{
.terminalStateFromMarkovChain
}
\keyword{internal}
