% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.check.R
\name{.type.check}
\alias{.type.check}
\title{.type_check}
\usage{
.type.check(input, must)
}
\arguments{
\item{must}{the type we need}

\item{check}{the input need to be check}
}
\description{
check whether the input has the correct type
}
