% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_probability.R
\name{markov_probability}
\alias{markov_probability}
\title{markov_probability}
\usage{
markov_probability(
  milo,
  diffusionmap,
  dpt,
  terminal_state,
  root_cell,
  scale_componet = TRUE,
  num_waypoints = 500
)
}
\arguments{
\item{milo}{milo or SingelCellExperiment object, used to store the result}

\item{diffusionmap}{DiffusionMap object}

\item{dpt}{DPT object}

\item{terminal_state}{}

\item{root_cell}{}

\item{num_waypoints}{integer
Number of waypoints to sample. 500L by default}

\item{scale_components}{logical
If True, the components will be scale}
}
\description{
Construct markov chain and calculate probabilities
}
