% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milo.umap.R
\name{milo.umap}
\alias{milo.umap}
\title{milo.umap}
\usage{
milo.umap(
  milo,
  slot_name = "UMAP_graph",
  n.neighbors = 50L,
  metric = "euclidean"
)
}
\arguments{
\item{milo}{the milo object need to conduct umap on}

\item{slot_name}{the slot name in reducedim where the result store}

\item{n.neighbors}{the number of neighboring points used}

\item{metric}{the choice of metric used to measure distance}
}
\value{
milo object with umap reduction
}
\description{
conduct the umap on the knn graph in umap
}
