% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milo_umap.R
\name{milo_umap}
\alias{milo_umap}
\title{milo_umap}
\usage{
milo_umap(
  milo,
  slot_name = "UMAP_knngraph",
  n.neighbors = 50L,
  metric = "euclidean"
)
}
\arguments{
\item{milo}{the milo object with knn graph that needed to conduct umap on.}

\item{slot_name}{character, with default 'UMAP_knngraph'.
\itemize{
\item The slot name in reduceDim where the result store
}}

\item{n.neighbors}{integer, with default 50L.
\itemize{
\item the number of neighboring points used
\item parameter of RunUMAP, checking its document for further detail
}}

\item{metric}{character, with default 'euclidean'
\itemize{
\item the choice of metric used to measure distance
\item parameter of RunUMAP, checking its document for further detail
}}
}
\value{
milo object with umap reduction
}
\description{
use function RunUMAP from Seurat to conduct the umap on the adjacency matrix of knn graph in milo object
}
\examples{

# load denpendency 
library(miloR)

# load example data
data(sce_vdj)
# get milo object with knn graph
sce_vdj <- setup_vdj_pseudobulk(sce_vdj, already.productive = FALSE)
traj_milo <- Milo(sce_vdj)
milo <- buildGraph(traj_milo, k = 50, d = 20, reduced.dim = "X_scvi")
milo <- makeNhoods(milo, reduced_dims = "X_scvi", d = 20)

# Construct UMAP
milo <- milo_umap(milo, n.neighbors = 10L, metric = "euclidean")

# visualize the result
scater::plotUMAP(milo,dimred = "UMAP_knngraph")

}
