% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miloUmap.R
\name{miloUmap}
\alias{miloUmap}
\title{Perform UMAP on the Adjacency Matrix of a Milo Object}
\usage{
miloUmap(
  milo,
  slot_name = "UMAP_knngraph",
  n.neighbors = 50L,
  metric = "euclidean"
)
}
\arguments{
\item{milo}{A \code{Milo} object containing a KNN graph. UMAP will be conducted on
the adjacency matrix of this graph.}

\item{slot_name}{Character. The name of the slot in \code{reducedDims} where the UMAP
results will be stored. Default is \code{'UMAP_knngraph'}.}

\item{n.neighbors}{Integer. The number of neighboring points used in the UMAP
algorithm. Default is \code{50L}. For further details, refer to the \code{RunUMAP}
documentation.}

\item{metric}{Character. The distance metric to be used in the UMAP algorithm.
Default is \code{'euclidean'}. For further details, refer to the \code{RunUMAP} documentation.}
}
\value{
milo object with umap reduction
}
\description{
This function uses \code{RunUMAP} from the Seurat package to perform UMAP dimensionality
reduction on the adjacency matrix of the KNN graph in a Milo object.
}
\examples{
data(sce_vdj)
sce_vdj <- setupVdjPseudobulk(sce_vdj,
    already.productive = FALSE
)
# Build Milo Object
traj_milo <- miloR::Milo(sce_vdj)
milo_object <- miloR::buildGraph(traj_milo, k = 50, d = 20, reduced.dim = "X_scvi")
milo_object <- miloR::makeNhoods(milo_object, reduced_dims = "X_scvi", d = 20)

# Construct UMAP on Milo Neighbor Graph
milo_object <- miloUmap(milo_object)

}
