% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pseudotime_to_cell.R
\name{project_pseudotime_to_cell}
\alias{project_pseudotime_to_cell}
\title{project_pseudotime_to_cell}
\usage{
project_pseudotime_to_cell(adata, pb_adata, term_states, suffix = "")
}
\arguments{
\item{adata}{SingleCellExperiment object}

\item{pb_adata}{pseudobulk data}

\item{term_states}{vector of terminal states with branch_probabilities to be transferred}

\item{suffix}{suffix to be added after the added column names, default ''}
}
\value{
subset of adata where cell that do not belong to any neighbourhood are removed and projected pseudotime information stored colData
}
\description{
Function to project pseudotime & branch probabilities from pb.milo (pseudobulk) to adata (cell).
}
\examples{
sce_vdj <- setup_vdj_pseudobulk(sce_vdj, 
                                already.productive = FALSE)
# Build Milo Object                                 
traj_milo <- miloR::Milo(sce_vdj)
milo_object <- miloR::buildGraph(traj_milo, k = 50, d = 20, reduced.dim = "X_scvi")
milo_object <- miloR::makeNhoods(milo_object, reduced_dims = "X_scvi", d = 20)

# Construct Pseudobulked VDJ Feature Space
pb.milo <- vdj_pseudobulk(milo_object, col_to_take = "anno_lvl_2_final_clean")
pbs = milo_object@nhoods
pb.milo <- runPCA(pb.milo, assay.type = "X")

# Define root and branch tips
pca <- t(as.matrix(reducedDim(pb.milo, type = "PCA")))
branch.tips <- c(which.max(pca[2, ]), which.min(pca[2, ]))
names(branch.tips) <- c("CD8+T", "CD4+T")
root <- which.min(pca[1, ])

# Construct Diffusion Map
dm <- DiffusionMap(t(pca), n_pcs = 50, n_eigs = 10)

#Markov Chain Construction 
pb.milo <- markov_probability(milo = pb.milo, 
                              diffusionmap = dm, 
                              terminal_state = branch.tips, 
                              root_cell = root, 
                              pseudotime_key = "pseudotime")

# Project Pseudobulk Data
cdata <- project_pseudotime_to_cell(milo_object, pb.milo, branch.tips)

}
