% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectProbability.R
\name{projectProbability}
\alias{projectProbability}
\title{Project Probabilities from Markov Chain to Pseudobulks}
\usage{
projectProbability(
  diffusionmap,
  waypoints,
  probabilities,
  t = 1,
  verbose = TRUE
)
}
\arguments{
\item{diffusionmap}{diffusion map, used to reconstruct diffustion distance matrix}

\item{waypoints}{Integer vector. Indices of the waypoints used in the Markov chain.}

\item{probabilities}{Numeric vector. Probabilities associated with the waypoints, calculated from the Markov chain.}

\item{t}{Numeric. The diffusion time to be used in the projection.}

\item{verbose}{Boolean, whether to print messages/warnings.}
}
\value{
each pseudobulk's probabilites
}
\description{
This function projects probabilities calculated from a Markov chain onto each pseudobulk based on a diffusion distance matrix.
}
