% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPseudotimeToCell.R
\name{projectPseudotimeToCell}
\alias{projectPseudotimeToCell}
\title{projectPseudotimeToCell}
\usage{
projectPseudotimeToCell(milo, pb_milo, term_states, suffix = "")
}
\arguments{
\item{milo}{SingleCellExperiment or milo object}

\item{pb_milo}{pseudobulk data}

\item{term_states}{vector of terminal states with branch_probabilities to be transferred}

\item{suffix}{suffix to be added after the added column names, default ''}
}
\value{
subset of milo or SingleCellExperiment object where cell that do not belong to any neighbourhood are removed and projected pseudotime information stored colData
}
\description{
Function to project pseudotime & branch probabilities from pb.milo (pseudobulk) to milo (cell).
}
\examples{
sce_vdj <- setupVdjPseudobulk(sce_vdj, 
                                already.productive = FALSE)
# Build Milo Object     
set.seed(100)                            
traj_milo <- miloR::Milo(sce_vdj)
milo_object <- miloR::buildGraph(traj_milo, k = 50, d = 20, reduced.dim = "X_scvi")
milo_object <- miloR::makeNhoods(milo_object, reduced_dims = "X_scvi", d = 20)

# Construct Pseudobulked VDJ Feature Space
pb.milo <- vdjPseudobulk(milo_object, col_to_take = "anno_lvl_2_final_clean")
pbs = milo_object@nhoods
pb.milo <- scater::runPCA(pb.milo, assay.type = "Feature_space")

# Define root and branch tips
pca <- t(as.matrix(SingleCellExperiment::reducedDim(pb.milo, type = "PCA")))
branch.tips <- c(232, 298)
names(branch.tips) <- c("CD8+T", "CD4+T")
root <- 476

# Construct Diffusion Map
dm <- destiny::DiffusionMap(t(pca), n_pcs = 50, n_eigs = 10)
dif.pse <- destiny::DPT(dm, tips = c(root, branch.tips), w_width = 0.1)

#Markov Chain Construction 
pb.milo <- markovProbability(milo = pb.milo, 
                              diffusionmap = dm, 
                              terminal_state = branch.tips, 
                              diffusiontime = dif.pse[[paste0("DPT", root)]],
                              root_cell = root, 
                              pseudotime_key = "pseudotime")

# Project Pseudobulk Data
projected_milo <- projectPseudotimeToCell(milo_object, pb.milo, branch.tips)

}
