% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_vdj_pseudobulk.R
\name{setup_vdj_pseudobulk}
\alias{setup_vdj_pseudobulk}
\title{setup_vdj_pseudobulk}
\usage{
setup_vdj_pseudobulk(
  sce,
  mode_option = c("abT", "gdT", "B"),
  productive_cols = NULL,
  productive_vdj = TRUE,
  productive_vj = TRUE,
  allowed_chain_status = c("Single pair", "Extra pair", "Extra pair-exception",
    "Orphan VDJ", "Orphan VDJ-exception"),
  subsetby = NULL,
  groups = NULL,
  extract_cols = NULL,
  filter_pattern = ",|None|No_cotig",
  check_vdj_mapping = c("v_call", "j_call"),
  check_vj_mapping = c("v_call", "j_call"),
  check_extract_cols_mapping = NULL,
  remove_missing = TRUE,
  already.productive = TRUE,
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object,
vdj data should be contained in colData for filtering}

\item{mode_option}{optional,
mode for extraction the V(D)J genes.}

\item{productive_cols}{character vector, names of colData used in productivity filtering, NULL by default}

\item{productive_vdj}{logical, TRUE by default
Option in productivity filtering. If True, cell will only be kept if the main VDJ chain is productive}

\item{productive_vj}{logical, TRUE by default
Option in productivity filtering. If True, cell will only be kept if the main VJ chain is productive}

\item{allowed_chain_status}{character vectors, optional, if specified,
the element should come from c("single pair","Extra pair", "Extra pair-exceptipn", "Orphan VDJ","Orphan VDJ-exception")}

\item{subsetby}{character, NULL by default, name of one colData provided for subsetting.}

\item{groups}{character vector, NULL by default, condition for subsetting}

\item{extract_cols}{character vector, NULL by default
column names where VDJ/VJ information is stored so that this will be used instead of the standard columns}

\item{filter_pattern}{character string, optional ",|None|No_contig" by default
pattern to filter form object. If NULL, does not filter}

\item{check_vdj_mapping}{character vector, optional
elements should come from c("v_call", "j_call")
Only columns in the argument will be checked for unclear mapping(contain comma) in VDJ calls}

\item{check_vj_mapping}{character vector, optional
elements should come from c("v_call", "j_call")
Only columns in the argument will be checked for unclear mapping(contain comma) in VJ calls}

\item{check_extract_cols_mapping}{character vecter, NULL by default
Only columns in the argument will be checked for unclear mapping (containing comma) in columns specified in extract_cols}

\item{remove_missing}{bool, True by default
If true, will remove cells with contigs matching the filter from the object. If False, will mask them with a uniform value dependent on the column name.}

\item{already.productive}{logical, whether the data is already filtered by productivity.}
}
\value{
filtered SingleCellExperiment object
}
\description{
Function for data preprocessing, filtering the data based on productivity, chain status, subset data, extract main TCR, and remove the unmapping data
}
\details{
data will undergo several process, including productivity filtering, chain status filtering, subseting, main TCR extracting, and unmapping data removing.
\itemize{
\item parameters for \strong{productivity filtering}:
\itemize{
\item already.productive
\itemize{
\item If true, the function will skip the productivity filtering
\item If false, retain only productive entries in columns specified by mode_option or productive_cols
}
\item mode_option
\itemize{
\item function will check colData(s) named \verb{productive_\{mode_option\}_\{type\}}, where type should be "VDJ" or "VJ" or both, depending on values of productive_vj and productive_vdj.
\item If set as "NULL", the function needs the option "extract_cols" to be specified
}
\item productive_cols
\itemize{
\item must be be specified when productivity filtering is need to conduct and mode_option is NULL.
\item where VDJ/VJ information is stored so that this will be used instead of the standard columns.
}
\item productive_vj, productive_vdj
\itemize{
\item If True, cell will only be kept if the main V(D)J chain is productive
}
}
\item parameter for \strong{chain status filtering}: allowed_chain_status, chain status to be kept
\item parameters for \strong{subsetting}: subsetby, groups
\itemize{
\item subsetting process will only be conducted when both parameters are provided. After subsetting, only the cell with {groups} feature in {subsetby} will be used for computing the VDJ feature space
}
}
}
