% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitCTgene.R
\name{splitCTgene}
\alias{splitCTgene}
\title{Split the V(D)J genes from \code{CTgene} column and store them separately.}
\usage{
splitCTgene(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object after conducting scRepertoire::combineTCR()}
}
\value{
list contain vector of VJ & VDJ genes from each cell
}
\description{
Split the V(D)J genes from \code{CTgene} column and store them separately.
}
\keyword{internal}
